+++ pacote.js
@@ -1,0 +1,4 @@
+
+export * from "constantes.js";
+export * from "funcoes.js";
+export * from "predefinido.js";
